+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Better support for schemas
 - Added `--all-schemas` option
-- Added `--schema` option to sync with schema
+- Added `--schema-first` option to sync with schema
 - Fixed issue with non-lowercase tables and partial syncs
 
 # 0.3.8